Reengineer parts of the Storage to make it more consistent (#210)
This included renaming methods to better suit the concept e.g. `Load`
and `Save` instead of `Dump`.

A small highlight is the `interface` based distinction between writeable
and readonly backends. Which led to more renaming. It also got rid of
the distinction between `loader` and `dumper` which are now handeld
through a single backend.

It also makes the API a bit more consice as do not longer need to know
that the loader pass into the constructer become the dumper.

Now you can control whether a `backend` is able to recieve writes by
implementing `Save` and thus conforming to the `WriteableBackend`
`interface`.