Allow operator goroutines to finish.
This was rather hard to track down, but what I noticed
was hanging/leaking goroutines waiting on `<-p.buf`. This came up during hunting
race conditions.

What happened here was that the gorouting that executed `opFunc` could
never return because the operator was somewhere trying to `Pull`. Which
in turn meant waiting on a message recievied `<-p.buf`. Altough we have
called `Stop` on the operator.

This fix now closes all ports resulting in more leaked goroutines.